requirements.txt
# Core dependencies
torch==2.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
requests==2.31.0
PyPDF2==3.0.1

# Transformers library and its ecosystem
transformers==4.37.2
tokenizers==0.14.1
filelock==3.13.1
huggingface-hub==0.20.3
packaging==23.2
tqdm==4.66.1
safetensors==0.4.2
regex==2023.12.25

# bitsandbytes for quantization (for supported platforms)
bitsandbytes==0.41.2

# Faiss can use scipy in some workflows
scipy==1.11.4

# Requests sub-dependencies (usually installed automatically, but explicit for reliability)
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
certifi==2024.2.2
